project/requirements.txt
aiosqlite==0.15.0
appdirs==1.4.4
asyncpg==0.21.0
atlastk==0.12
attrs==20.2.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.3
fastapi==0.61.1
flake8==3.8.3
gunicorn==20.0.4
h11==0.10.0
httptools==0.1.1
idna==2.10
iniconfig==1.0.1
iso8601==0.1.13
isort==5.5.4
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pydantic==1.6.1
pyflakes==2.2.0
pyparsing==2.4.7
PyPika==0.42.1
pytest==6.1.0
pytest-cov==2.10.1
regex==2020.9.27
requests==2.24.0
six==1.15.0
starlette==0.13.6
toml==0.10.1
tortoise-orm==0.16.16
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.12.0
uvloop==0.14.0
